requirements.txt
# Core Flask dependencies
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.5
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0

# Flask extensions
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Cors==5.0.0

# Database
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.14.0

# Forms and validation
WTForms==3.1.2
email-validator==2.1.1

# Image handling
cloudinary==1.39.0
Pillow==11.1.0

# Security
bcrypt==4.2.0
PyJWT==2.10.1

# Utilities
python-dotenv==1.0.1
gunicorn==22.0.0

# Date and time
pytz==2025.1
